backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.3
mangum==0.17.0
